First move the code into a new file
Make sure correct formatting
