+++ Refit/RequestBuilderImplementation.cs
@@ -288,142 +288,4 @@ namespace Refit
             block();
         }
     }
-
-    public class RestMethodInfo
-    {
-        public string Name { get; set; }
-        public Type Type { get; set; }
-        public MethodInfo MethodInfo { get; set; }
-        public HttpMethod HttpMethod { get; set; }
-        public string RelativePath { get; set; }
-        public Dictionary<int, string> ParameterMap { get; set; }
-        public Tuple<BodySerializationMethod, int> BodyParameterInfo { get; set; }
-        public Dictionary<int, string> QueryParameterMap { get; set; }
-        public Type ReturnType { get; set; }
-        public Type SerializedReturnType { get; set; }
-
-        static readonly Regex parameterRegex = new Regex(@"^{(.*)}$");
-
-        public RestMethodInfo(Type targetInterface, MethodInfo methodInfo)
-        {
-            Type = targetInterface;
-            Name = methodInfo.Name;
-            MethodInfo = methodInfo;
-
-            var hma = methodInfo.GetCustomAttributes(true)
-                .OfType<HttpMethodAttribute>()
-                .First();
-
-            HttpMethod = hma.Method;
-            RelativePath = hma.Path;
-
-            verifyUrlPathIsSane(RelativePath);
-            determineReturnTypeInfo(methodInfo);
-
-            var parameterList = methodInfo.GetParameters().ToList();
-
-            ParameterMap = buildParameterMap(RelativePath, parameterList);
-            BodyParameterInfo = findBodyParameter(parameterList);
-
-            QueryParameterMap = new Dictionary<int, string>();
-            for (int i=0; i < parameterList.Count; i++) {
-                if (ParameterMap.ContainsKey(i) || (BodyParameterInfo != null && BodyParameterInfo.Item2 == i)) {
-                    continue;
-                }
-
-                QueryParameterMap[i] = getUrlNameForParameter(parameterList[i]);
-            }
-        }
-
-        void verifyUrlPathIsSane(string relativePath)
-        {
-            if (!relativePath.StartsWith("/")) {
-                goto bogusPath;
-            }
-
-            var parts = relativePath.Split('/');
-            if (parts.Length == 0) {
-                goto bogusPath;
-            }
-
-            return;
-
-        bogusPath:
-            throw new ArgumentException("URL path must be of the form '/foo/bar/baz'");
-        }
-
-        Dictionary<int, string> buildParameterMap(string relativePath, List<ParameterInfo> parameterInfo)
-        {
-            var ret = new Dictionary<int, string>();
-
-            var parameterizedParts = relativePath.Split('/', '?').SelectMany(x => {
-                var m = parameterRegex.Match(x);
-                return (m.Success ? EnumerableEx.Return(m) : Enumerable.Empty<Match>());
-            }).ToList();
-
-            if (parameterizedParts.Count == 0) {
-                return ret;
-            }
-
-            var paramValidationDict = parameterInfo.ToDictionary(k => getUrlNameForParameter(k).ToLowerInvariant(), v => v);
-
-            foreach (var match in parameterizedParts) {
-                var name = match.Groups[1].Value.ToLowerInvariant();
-                if (!paramValidationDict.ContainsKey(name)) {
-                    throw new ArgumentException(String.Format("URL has parameter {0}, but no method parameter matches", name));
-                }
-
-                ret.Add(parameterInfo.IndexOf(paramValidationDict[name]), name);
-            }
-
-            return ret;
-        }
-
-        string getUrlNameForParameter(ParameterInfo paramInfo)
-        {
-            var aliasAttr = paramInfo.GetCustomAttributes(true)
-                .OfType<AliasAsAttribute>()
-                .FirstOrDefault();
-            return aliasAttr != null ? aliasAttr.Name : paramInfo.Name;
-        }
-
-        Tuple<BodySerializationMethod, int> findBodyParameter(List<ParameterInfo> parameterList)
-        {
-            var bodyParams = parameterList
-                .Select(x => new { Parameter = x, BodyAttribute = x.GetCustomAttributes(true).OfType<BodyAttribute>().FirstOrDefault() })
-                .Where(x => x.BodyAttribute != null)
-                .ToList();
-
-            if (bodyParams.Count > 1) {
-                throw new ArgumentException("Only one parameter can be a Body parameter");
-            }
-
-            if (bodyParams.Count == 0) {
-                return null;
-            }
-
-            var ret = bodyParams[0];
-            return Tuple.Create(ret.BodyAttribute.SerializationMethod, parameterList.IndexOf(ret.Parameter));
-        }
-
-        void determineReturnTypeInfo(MethodInfo methodInfo)
-        {
-            if (methodInfo.ReturnType.IsGenericType == false && methodInfo.ReturnType != typeof(Task)) {
-                goto bogusMethod;
-            }
-
-            var genericType = methodInfo.ReturnType.GetGenericTypeDefinition();
-            if (genericType != typeof(Task<>) && genericType != typeof(IObservable<>)) {
-                goto bogusMethod;
-            }
-
-            ReturnType = methodInfo.ReturnType;
-            SerializedReturnType = methodInfo.ReturnType.GetGenericArguments()[0];
-            if (SerializedReturnType == typeof(HttpResponseMessage)) SerializedReturnType = null;
-            return;
-
-        bogusMethod:
-            throw new ArgumentException("All REST Methods must return either Task<T> or IObservable<T>");
-        }
-    }
 }
